Update and rename intensity-based_segmentation.py to intensity_based_segmentation.py
